# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ComptaAfrica API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  totalRevenue: number;
  totalExpenses: number;
  netProfit: number;
  revenueGrowth: number;
  expensesGrowth: number;
  profitGrowth: number;
  pendingInvoices: number;
  totalClients: number;
}

export interface ChartDataPoint {
  month: string;
  revenue: number;
  expenses: number;
  profit: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  transaction: "transaction",
  invoice: "invoice",
  client: "client",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  amount?: number | null;
  date: string;
  category?: string | null;
}

export type AlertType = (typeof AlertType)[keyof typeof AlertType];

export const AlertType = {
  warning: "warning",
  info: "info",
  success: "success",
  error: "error",
} as const;

export interface Alert {
  id: number;
  type: AlertType;
  title: string;
  message: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  revenue: "revenue",
  expense: "expense",
} as const;

export interface Transaction {
  id: number;
  date: string;
  description: string;
  amount: number;
  type: TransactionType;
  category: string;
  clientId?: number | null;
  clientName?: string | null;
}

export type CreateTransactionBodyType =
  (typeof CreateTransactionBodyType)[keyof typeof CreateTransactionBodyType];

export const CreateTransactionBodyType = {
  revenue: "revenue",
  expense: "expense",
} as const;

export interface CreateTransactionBody {
  date: string;
  description: string;
  amount: number;
  type: CreateTransactionBodyType;
  category: string;
  clientId?: number | null;
}

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus];

export const InvoiceStatus = {
  draft: "draft",
  sent: "sent",
  paid: "paid",
  overdue: "overdue",
} as const;

export interface InvoiceItem {
  description: string;
  quantity: number;
  unitPrice: number;
  total: number;
}

export interface Invoice {
  id: number;
  number: string;
  clientId: number;
  clientName: string;
  date: string;
  dueDate: string;
  status: InvoiceStatus;
  total: number;
  items: InvoiceItem[];
}

export interface CreateInvoiceBody {
  clientId: number;
  date: string;
  dueDate: string;
  items: InvoiceItem[];
}

export type UpdateInvoiceBodyStatus =
  (typeof UpdateInvoiceBodyStatus)[keyof typeof UpdateInvoiceBodyStatus];

export const UpdateInvoiceBodyStatus = {
  draft: "draft",
  sent: "sent",
  paid: "paid",
  overdue: "overdue",
} as const;

export interface UpdateInvoiceBody {
  status?: UpdateInvoiceBodyStatus;
}

export interface Client {
  id: number;
  name: string;
  email: string;
  phone: string;
  address?: string;
  totalPaid: number;
  transactionCount: number;
  createdAt: string;
}

export type ClientDetail = Client & {
  transactions: Transaction[];
};

export interface CreateClientBody {
  name: string;
  email: string;
  phone: string;
  address?: string;
}

export type ExpenseCategory =
  (typeof ExpenseCategory)[keyof typeof ExpenseCategory];

export const ExpenseCategory = {
  transport: "transport",
  stock: "stock",
  loyer: "loyer",
  salaires: "salaires",
  marketing: "marketing",
  utilitaires: "utilitaires",
  autre: "autre",
} as const;

export interface Expense {
  id: number;
  date: string;
  description: string;
  amount: number;
  category: ExpenseCategory;
  receipt?: string | null;
}

export type CreateExpenseBodyCategory =
  (typeof CreateExpenseBodyCategory)[keyof typeof CreateExpenseBodyCategory];

export const CreateExpenseBodyCategory = {
  transport: "transport",
  stock: "stock",
  loyer: "loyer",
  salaires: "salaires",
  marketing: "marketing",
  utilitaires: "utilitaires",
  autre: "autre",
} as const;

export interface CreateExpenseBody {
  date: string;
  description: string;
  amount: number;
  category: CreateExpenseBodyCategory;
}

export interface ExpenseCategoryTotal {
  category: string;
  total: number;
  percentage: number;
  count: number;
}

export interface MonthlyReport {
  month: string;
  revenue: number;
  expenses: number;
  profit: number;
  transactionCount: number;
}

export interface ReportSummary {
  totalRevenue: number;
  totalExpenses: number;
  netProfit: number;
  profitMargin: number;
  bestMonth: string;
  worstMonth: string;
}

export type GetDashboardChartParams = {
  year?: number;
};

export type ListTransactionsParams = {
  type?: ListTransactionsType;
  from?: string;
  to?: string;
  search?: string;
};

export type ListTransactionsType =
  (typeof ListTransactionsType)[keyof typeof ListTransactionsType];

export const ListTransactionsType = {
  revenue: "revenue",
  expense: "expense",
} as const;

export type ListInvoicesParams = {
  status?: ListInvoicesStatus;
};

export type ListInvoicesStatus =
  (typeof ListInvoicesStatus)[keyof typeof ListInvoicesStatus];

export const ListInvoicesStatus = {
  draft: "draft",
  sent: "sent",
  paid: "paid",
  overdue: "overdue",
} as const;

export type ListClientsParams = {
  search?: string;
};

export type ListExpensesParams = {
  category?: string;
};

export type GetMonthlyReportParams = {
  year?: number;
};