# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ComptaAfrica API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Dashboard summary and analytics
  - name: transactions
    description: Financial transactions
  - name: invoices
    description: Invoice management
  - name: clients
    description: Client management
  - name: expenses
    description: Expense management
  - name: reports
    description: Financial reports

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary with KPIs
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/chart:
    get:
      operationId: getDashboardChart
      tags: [dashboard]
      summary: Get revenue vs expenses chart data (monthly)
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Chart data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChartDataPoint"

  /dashboard/recent-activity:
    get:
      operationId: getDashboardRecentActivity
      tags: [dashboard]
      summary: Get recent transactions and activities
      responses:
        "200":
          description: Recent activity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  /dashboard/alerts:
    get:
      operationId: getDashboardAlerts
      tags: [dashboard]
      summary: Get smart alerts for dashboard
      responses:
        "200":
          description: Alert list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"

  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [revenue, expense]
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, sent, paid, overdue]
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create a new invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceBody"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Client list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID with history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client with transaction history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"

  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List all expenses
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Expense list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create a new expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"

  /expenses/by-category:
    get:
      operationId: getExpensesByCategory
      tags: [expenses]
      summary: Get expenses grouped by category
      responses:
        "200":
          description: Expenses by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExpenseCategoryTotal"

  /reports/monthly:
    get:
      operationId: getMonthlyReport
      tags: [reports]
      summary: Get monthly financial report
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyReport"

  /reports/summary:
    get:
      operationId: getReportSummary
      tags: [reports]
      summary: Get overall financial summary
      responses:
        "200":
          description: Financial summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    DashboardSummary:
      type: object
      properties:
        totalRevenue:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        revenueGrowth:
          type: number
        expensesGrowth:
          type: number
        profitGrowth:
          type: number
        pendingInvoices:
          type: integer
        totalClients:
          type: integer
      required:
        - totalRevenue
        - totalExpenses
        - netProfit
        - revenueGrowth
        - expensesGrowth
        - profitGrowth
        - pendingInvoices
        - totalClients

    ChartDataPoint:
      type: object
      properties:
        month:
          type: string
        revenue:
          type: number
        expenses:
          type: number
        profit:
          type: number
      required:
        - month
        - revenue
        - expenses
        - profit

    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [transaction, invoice, client]
        description:
          type: string
        amount:
          type: number
          nullable: true
        date:
          type: string
          format: date-time
        category:
          type: string
          nullable: true
      required:
        - id
        - type
        - description
        - date

    Alert:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [warning, info, success, error]
        title:
          type: string
        message:
          type: string
      required:
        - id
        - type
        - title
        - message

    Transaction:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [revenue, expense]
        category:
          type: string
        clientId:
          type: integer
          nullable: true
        clientName:
          type: string
          nullable: true
      required:
        - id
        - date
        - description
        - amount
        - type
        - category

    CreateTransactionBody:
      type: object
      properties:
        date:
          type: string
          format: date
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [revenue, expense]
        category:
          type: string
        clientId:
          type: integer
          nullable: true
      required:
        - date
        - description
        - amount
        - type
        - category

    Invoice:
      type: object
      properties:
        id:
          type: integer
        number:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        date:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum: [draft, sent, paid, overdue]
        total:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
      required:
        - id
        - number
        - clientId
        - clientName
        - date
        - dueDate
        - status
        - total
        - items

    InvoiceItem:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
      required:
        - description
        - quantity
        - unitPrice
        - total

    CreateInvoiceBody:
      type: object
      properties:
        clientId:
          type: integer
        date:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
      required:
        - clientId
        - date
        - dueDate
        - items

    UpdateInvoiceBody:
      type: object
      properties:
        status:
          type: string
          enum: [draft, sent, paid, overdue]

    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        totalPaid:
          type: number
        transactionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - phone
        - totalPaid
        - transactionCount
        - createdAt

    ClientDetail:
      allOf:
        - $ref: "#/components/schemas/Client"
        - type: object
          properties:
            transactions:
              type: array
              items:
                $ref: "#/components/schemas/Transaction"
          required:
            - transactions

    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
      required:
        - name
        - email
        - phone

    Expense:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        description:
          type: string
        amount:
          type: number
        category:
          type: string
          enum: [transport, stock, loyer, salaires, marketing, utilitaires, autre]
        receipt:
          type: string
          nullable: true
      required:
        - id
        - date
        - description
        - amount
        - category

    CreateExpenseBody:
      type: object
      properties:
        date:
          type: string
          format: date
        description:
          type: string
        amount:
          type: number
        category:
          type: string
          enum: [transport, stock, loyer, salaires, marketing, utilitaires, autre]
      required:
        - date
        - description
        - amount
        - category

    ExpenseCategoryTotal:
      type: object
      properties:
        category:
          type: string
        total:
          type: number
        percentage:
          type: number
        count:
          type: integer
      required:
        - category
        - total
        - percentage
        - count

    MonthlyReport:
      type: object
      properties:
        month:
          type: string
        revenue:
          type: number
        expenses:
          type: number
        profit:
          type: number
        transactionCount:
          type: integer
      required:
        - month
        - revenue
        - expenses
        - profit
        - transactionCount

    ReportSummary:
      type: object
      properties:
        totalRevenue:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        profitMargin:
          type: number
        bestMonth:
          type: string
        worstMonth:
          type: string
      required:
        - totalRevenue
        - totalExpenses
        - netProfit
        - profitMargin
        - bestMonth
        - worstMonth