# lib/api-zod/src/generated/types/createExpenseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ComptaAfrica API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateExpenseBodyCategory } from "./createExpenseBodyCategory";

export interface CreateExpenseBody {
  date: Date;
  description: string;
  amount: number;
  category: CreateExpenseBodyCategory;
}