# lib/api-zod/src/generated/types/createInvoiceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ComptaAfrica API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InvoiceItem } from "./invoiceItem";

export interface CreateInvoiceBody {
  clientId: number;
  date: Date;
  dueDate: Date;
  items: InvoiceItem[];
}